requirements.txt
asn1crypto==0.22.0
Augmentor==0.1.8
bleach==1.5.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
conda==4.3.27
cryptography==2.0.3
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
Flask==0.12.2
future==0.16.0
html5lib==0.9999999
idna==2.6
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.1
Markdown==2.6.9
MarkupSafe==1.0
mistune==0.8.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.13.3
olefile==0.44
pandas==0.21.0
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
ptyprocess==0.5.2
pycosat==0.6.2
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==17.2.0
PySocks==1.6.7
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
ruamel-yaml==0.11.14
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc3
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
tqdm==4.19.4
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.0.8
yapf==0.19.0
